Load rc-tooltip dynamically
